Refactor entry point to support web server interfacing; feature queries expanded; removed demo data; support added for ingesting tsv expression matrix
