+++ requirements.txt
@@ -7,3 +7,9 @@ bravado-core==5.10.2
 pyyaml>=4.2b1
 h5py==2.9.0
 swagger-ui-bundle==0.0.3
+pandas==0.24.2
+
+# jsonschema[format] requires the following otherwise we get an error:
+rfc3987==1.3.8
+strict-rfc3339==0.7
+webcolors==1.8.1
